disable all actions until multi-filename operations are implemented
It's better to fail loudly and do nothing than to introduce possibly
fatal errors during linking or deleting by ignoring some instances of
hardlinked files.

Use oom() for the errors because of least coding effort - this is a
temporary change anyways.

The error will be removed from an action as soon as multi-filename
support is available for that action.